Pull request review changes - removed unnecessary imports - restructured the project:   - pages folder with main pages   - conf folder with configuration files and interfaces   - dynamic color and type for animal icons in CardGrid.tsx
