+++ app/src/conf/animalIcons.ts
@@ -1,0 +1,21 @@
+export interface AnimalIconsInterface {
+  crocodile: {
+    color: string;
+    name: string;
+  };
+  cat: {
+    color: string;
+    name: string;
+  };
+}
+
+export const animalIcons: AnimalIconsInterface = {
+  crocodile: {
+    color: "#7bb853",
+    name: "alligator",
+  },
+  cat: {
+    color: "#f0cc55",
+    name: "tiger",
+  },
+};
