+++ app/src/conf/animals.ts
@@ -1,0 +1,9 @@
+export interface AnimalInterface {
+  type: string;
+  animal: string;
+  name: string;
+  img: string;
+  description: string;
+  size: number;
+  weigth: number;
+}
