+++ app/src/index.scss
@@ -36,5 +36,5 @@ body {
   -moz-osx-font-smoothing: grayscale;
 }
 
-@import "./components/dashboard.scss";
+@import "./pages/dashboard.scss";
 @import "./components/cardGrid.scss";
