+++ app/src/pages/AboutPage.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 const About = () => {
   return (
     <div>
