+++ app/src/pages/ContactPage.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 const Contact = () => {
   return (
     <div>
