+++ app/src/pages/OopsiePage.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 const Oopsie = () => {
   return (
     <div>
