+++ app/src/pages/dashboard.scss
@@ -16,6 +16,7 @@
     a {
       padding: 1rem 0 1rem 2rem;
       text-decoration: none;
+      color: $violet;
 
       &:first-of-type {
         padding-left: 0;
